[MOV] website, html_builder: move image and background options
As mass_mailing will soon use the `html_builder` instead of the legacy editor,
some useful options currently exclusive to `website` can be moved to be globally
available.

task-4247642
