+++ addons/html_builder/static/src/plugins/background_option/background_image.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates xml:space="preserve">
 
-<t t-name="website.BackgroundImageOption">
+<t t-name="html_builder.BackgroundImageOption">
     <BuilderRow t-if="isActiveItem('toggle_bg_image_id')" label.translate="Image" level="1">
         <BuilderButton preview="false" title.translate="Edit image" type="'success'" action="'replaceBgImage'"> Replace </BuilderButton>
         <ImageSize />
