+++ addons/html_builder/static/src/plugins/background_option/background_option.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates xml:space="preserve">
 
-<t t-name="website.BackgroundOption">
+<t t-name="html_builder.BackgroundOption">
     <!-- BackgroundToggler -->
     <BuilderRow label.translate="Background" t-if="props.withColors or props.withImages">
     <!-- todo adapt when colorpicker is implemented: snippet_options_background_color_widget-->
