+++ addons/html_builder/static/src/plugins/background_option/background_position_option.js
@@ -1,6 +1,6 @@
 import { BaseOptionComponent } from "@html_builder/core/utils";
 
 export class BackgroundPositionOption extends BaseOptionComponent {
-    static template = "website.BackgroundPositionOption";
+    static template = "html_builder.BackgroundPositionOption";
     static props = {};
 }
