+++ addons/html_builder/static/src/plugins/background_option/background_position_option.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates xml:space="preserve">
 
-<t t-name="website.BackgroundPositionOption">
+<t t-name="html_builder.BackgroundPositionOption">
     <t t-if="isActiveItem('toggle_bg_image_id')">
         <BuilderRow label.translate="Position" level="2">
             <BuilderSelect action="'backgroundType'" preview="false">
