+++ addons/html_builder/static/src/plugins/background_option/background_position_overlay.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates xml:space="preserve">
 
-<t t-name="website.BackgroundPositionOverlay">
+<t t-name="html_builder.BackgroundPositionOverlay">
     <div class="o_we_background_position_overlay" t-ref="backgroundOverlay">
         <div class="o_we_overlay_content position-absolute" t-ref="overlayContent">
             <div class="o_overlay_background" t-on-mousedown.prevent="onDragBackgroundStart"
