+++ addons/html_builder/static/src/plugins/background_option/background_shape_option_plugin.js
