+++ addons/html_builder/static/src/plugins/background_option/background_shapes_definition.js
