+++ addons/html_builder/static/src/plugins/image/image_size_plugin.js
